+++ built/GrammarLanguage.js
@@ -49,8 +49,6 @@ var GrammarConstants;
     GrammarConstants["baseNodeType"] = "baseNodeType";
     GrammarConstants["blobNode"] = "blobNode";
     GrammarConstants["errorNode"] = "errorNode";
-    GrammarConstants["terminalNode"] = "terminalNode";
-    GrammarConstants["nonTerminalNode"] = "nonTerminalNode";
     // parse time
     GrammarConstants["extends"] = "extends";
     GrammarConstants["abstract"] = "abstract";
@@ -326,7 +324,19 @@ class GrammarBackedNonRootNode extends GrammarBackedNode {
         return str ? TreeUtils_1.default.formatStr(str, catchAllCellDelimiter, this.cells) : this.getLine();
     }
     compile() {
-        return this._getCompiledIndentation() + this._getCompiledLine();
+        const def = this.getDefinition();
+        if (def.isTerminalNodeType())
+            return this._getCompiledIndentation() + this._getCompiledLine();
+        const compiler = def._getCompilerObject();
+        const openChildrenString = compiler[GrammarConstantsCompiler.openChildren] || "";
+        const closeChildrenString = compiler[GrammarConstantsCompiler.closeChildren] || "";
+        const childJoinCharacter = compiler[GrammarConstantsCompiler.joinChildrenWith] || "\n";
+        const compiledLine = this._getCompiledLine();
+        const indent = this._getCompiledIndentation();
+        const compiledChildren = this.map(child => child.compile()).join(childJoinCharacter);
+        return `${indent}${compiledLine}${openChildrenString}
+${compiledChildren}
+${indent}${closeChildrenString}`;
     }
     // todo: remove
     get cells() {
@@ -343,9 +353,7 @@ class GrammarBackedNonRootNode extends GrammarBackedNode {
         return cells;
     }
 }
-class GrammarBackedTerminalNode extends GrammarBackedNonRootNode {
-}
-exports.GrammarBackedTerminalNode = GrammarBackedTerminalNode;
+exports.GrammarBackedNonRootNode = GrammarBackedNonRootNode;
 class GrammarBackedErrorNode extends GrammarBackedNonRootNode {
     // todo: is this correct?
     getLineCellTypes() {
@@ -356,25 +364,6 @@ class GrammarBackedErrorNode extends GrammarBackedNonRootNode {
     }
 }
 exports.GrammarBackedErrorNode = GrammarBackedErrorNode;
-class GrammarBackedNonTerminalNode extends GrammarBackedNonRootNode {
-    // todo: implement
-    _getChildJoinCharacter() {
-        return "\n";
-    }
-    compile() {
-        const compiler = this.getDefinition()._getCompilerObject();
-        const openChildrenString = compiler[GrammarConstantsCompiler.openChildren] || "";
-        const closeChildrenString = compiler[GrammarConstantsCompiler.closeChildren] || "";
-        const childJoinCharacter = compiler[GrammarConstantsCompiler.joinChildrenWith] || "\n";
-        const compiledLine = this._getCompiledLine();
-        const indent = this._getCompiledIndentation();
-        const compiledChildren = this.map(child => child.compile()).join(childJoinCharacter);
-        return `${indent}${compiledLine}${openChildrenString}
-${compiledChildren}
-${indent}${closeChildrenString}`;
-    }
-}
-exports.GrammarBackedNonTerminalNode = GrammarBackedNonTerminalNode;
 class GrammarBackedBlobNode extends GrammarBackedNonRootNode {
     getFirstWordMap() {
         return {};
@@ -1199,7 +1188,11 @@ class AbstractGrammarDefinitionNode extends AbstractExtendibleTreeNode {
         const className = this._getGeneratedClassName();
         const extendsClassName = this._getExtendsClassName(false);
         const gettersAndConstants = this._getCellGettersAndNodeTypeConstants();
+        const langProgram = this.getLanguageDefinitionProgram();
+        const rootNode = langProgram._getRootNodeTypeDefinitionNode();
+        const amIRoot = rootNode === this;
         const code = `class ${className} extends ${extendsClassName} {
+      ${amIRoot ? `getGrammarProgramRoot() {return jtree.GrammarProgram._languages.${rootNode.getNodeTypeIdFromDefinition()} }` : ""}
       ${gettersAndConstants}
       ${this._getCustomJavascriptMethods()}
 }`;
@@ -1385,10 +1378,13 @@ class NonRootNodeTypeDefinition extends AbstractGrammarDefinitionNode {
         const chain = this._getNodeTypeInheritanceSet();
         return firstWordsInScope.some(firstWord => chain.has(firstWord));
     }
+    isTerminalNodeType() {
+        return !this._getFromExtended(GrammarConstants.inScope) && !this._getFromExtended(GrammarConstants.catchAllNodeType);
+    }
     _getExtendsClassName(isCompiled = false) {
         if (this._amIRoot())
             return "jtree.GrammarBackedRootNode";
-        const isNonTerminal = this._getFromExtended(GrammarConstants.inScope) || this._getFromExtended(GrammarConstants.catchAllNodeType);
+        const isNonTerminal = !this.isTerminalNodeType();
         const extendedDef = this._getExtendedParent();
         // todo: this is broken for runtime stuff.
         if (!isCompiled) {
@@ -1400,12 +1396,10 @@ class NonRootNodeTypeDefinition extends AbstractGrammarDefinitionNode {
             const baseTypeNames = {};
             baseTypeNames[GrammarConstants.blobNode] = "BlobNode";
             baseTypeNames[GrammarConstants.errorNode] = "ErrorNode";
-            baseTypeNames[GrammarConstants.terminalNode] = "TerminalNode";
-            baseTypeNames[GrammarConstants.nonTerminalNode] = "NonTerminalNode";
-            const baseNodeTypeName = baseTypeNames[this._getFromExtended(GrammarConstants.baseNodeType)] || (isNonTerminal ? "NonTerminalNode" : "TerminalNode");
+            const baseNodeTypeName = baseTypeNames[this._getFromExtended(GrammarConstants.baseNodeType)] || "GrammarBackedNonRootNode";
             return `jtree.` + baseNodeTypeName; // todo: this is incorrect but works for legacy stuff.
         }
-        return extendedDef ? extendedDef._getGeneratedClassName() : isNonTerminal ? "jtree.TerminalNode" : "jtree.NonTerminalNode";
+        return extendedDef ? extendedDef._getGeneratedClassName() : "jtree.GrammarBackedNonRootNode";
     }
     _getFirstCellHighlightScope() {
         const program = this.getLanguageDefinitionProgram();
@@ -1625,15 +1619,12 @@ class GrammarProgram extends AbstractGrammarDefinitionNode {
         return this._getRootNodeTypeDefinitionNode().get(GrammarConstants.catchAllNodeType);
     }
     _getRootConstructor() {
-        const extendedConstructor = this._getRootNodeTypeDefinitionNode()._getConstructorDefinedInGrammar();
-        const grammarProgram = this;
-        // Note: this is some of the most unorthodox code in this repo. We create a class on the fly for your
-        // new language.
-        return class extends extendedConstructor {
-            getGrammarProgramRoot() {
-                return grammarProgram;
-            }
-        };
+        const def = this._getRootNodeTypeDefinitionNode();
+        const langName = def.getNodeTypeIdFromDefinition();
+        // todo: check for namespace conflicts!
+        if (!GrammarProgram._languages[langName])
+            GrammarProgram._languages[langName] = this;
+        return def._getConstructorDefinedInGrammar();
     }
     getRootConstructor() {
         if (!this._cache_rootConstructorClass)
@@ -1732,4 +1723,5 @@ ${GrammarConstants.nodeType} anyNode
 ${GrammarConstants.cellType} anyWord`).getRootConstructor();
     }
 }
+GrammarProgram._languages = {};
 exports.GrammarProgram = GrammarProgram;