+++ built/jtree.d.ts
@@ -1,16 +1,15 @@
 import TreeUtils from "./base/TreeUtils";
 import TreeNode from "./base/TreeNode";
-import { GrammarProgram, GrammarBackedRootNode, GrammarBackedTerminalNode, GrammarBackedNonTerminalNode, GrammarBackedBlobNode, GrammarBackedErrorNode } from "./GrammarLanguage";
+import { GrammarProgram, GrammarBackedRootNode, GrammarBackedNonRootNode, GrammarBackedBlobNode, GrammarBackedErrorNode } from "./GrammarLanguage";
 import UnknownGrammarProgram from "./tools/UnknownGrammarProgram";
 import TreeNotationCodeMirrorMode from "./codemirror/TreeNotationCodeMirrorMode";
 declare class jtree {
     static GrammarBackedRootNode: typeof GrammarBackedRootNode;
+    static GrammarBackedNonRootNode: typeof GrammarBackedNonRootNode;
     static Utils: typeof TreeUtils;
     static TreeNode: typeof TreeNode;
-    static NonTerminalNode: typeof GrammarBackedNonTerminalNode;
     static BlobNode: typeof GrammarBackedBlobNode;
     static ErrorNode: typeof GrammarBackedErrorNode;
-    static TerminalNode: typeof GrammarBackedTerminalNode;
     static GrammarProgram: typeof GrammarProgram;
     static UnknownGrammarProgram: typeof UnknownGrammarProgram;
     static TreeNotationCodeMirrorMode: typeof TreeNotationCodeMirrorMode;